chore: move selectCalculatedFees to common
